Bug #451418, support for time stamp error clip(DTS)
[root cause  ] Presentation Time Stamp (PTS) is error
[changes     ] revise video decoder omx component
[side effects] NO
[self test   ]<board, Test cases summary>:Tsahrk, video
playback/thumbnail ok.
[reviewers   ] xiaowei.luo
[change_type ] AOB --> sprd_adaptation
[tag_product ] common

+++ sprd_codecs/openmax/avc_h264_mix/enc/hw/sc8830/include/sci_types.h
@@ -53,8 +53,8 @@ typedef unsigned char		BOOLEAN;
 typedef unsigned char		uint8;
 typedef unsigned short		uint16;
 typedef unsigned int		uint32;
-typedef unsigned long int 	uint64;
-typedef long int 			int64;
+typedef unsigned long long 	uint64;
+typedef signed long long	int64;
 typedef unsigned int		uint;
 typedef unsigned long		uint_32or64;
